requirements.txt
# Python
python>=3.10

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# API & HTTP
requests==2.31.0
aiohttp==3.9.1
httpx==0.26.0

# HTML Parsing
beautifulsoup4==4.12.3
lxml==5.1.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Async
asyncio==3.4.3

# Environment
python-dotenv==1.0.0

# Logging
colorlog==6.8.2

# Utilities
click==8.1.7
tqdm==4.66.1
